Entity consolidation

+++ api/src/main/java/com/ruby/rubia_server/core/dto/UpdateMessageDTO.java
@@ -24,5 +24,5 @@ public class UpdateMessageDTO {
     
     private Boolean isAiGenerated;
     
-    private BigDecimal aiConfidence;
+    private Double aiConfidence;
 }
